docs/storage_research/test_stand/requirements.txt
psycopg2-binary
cassandra-driver
clickhouse-driver
